+++ docs/changes.rst
@@ -2,12 +2,38 @@
 Changes
 =======
 
-v1.3.0 (UNRELEASED)
+v2.0.0 (UNRELEASED)
 ===================
 
 - Drop support for Python 2.6, 3.2, and 3.3.
 
 - Include CPython 3.5, 3.6 and 3.7, and PyPy 3.5 in the test matrix.
+
+- **Backwards incompatible:** :class:`pykka.Future.map` on a future with an
+  iterable result no longer applies the map function to each item in iterable.
+  Instead, the entire future result is passed to the map function. (Fixes:
+  :issue:`64`)
+
+  To upgrade existing code, make sure to explictly apply the core of your map
+  function to each item in the iterable::
+
+      >>> f = pykka.ThreadingFuture()
+      >>> f.set([1, 2, 3])
+      >>> f.map(lambda x: x + 1).get()  # Pykka < 2.0
+      [2, 3, 4]
+      >>> f.map(lambda x: [i + 1 for i in x]).get()  # Pykka >= 2.0
+      [2, 3, 4]
+
+  This change makes it easy to use :meth:`~pykka.Future.map` to extract a field
+  from a future that returns a dict::
+
+      >>> f = pykka.ThreadingFuture()
+      >>> f.set({'foo': 'bar'})
+      >>> f.map(lambda x: x['foo']).get()
+      'bar'
+
+  Because dict is an iterable, the now removed special handling of iterables
+  made this pattern difficult to use.
 
 - Include gevent and Eventlet tests in all environments. Since Pykka was
   originally developed, both has grown support for Python 3 and PyPy.